// src/Model/user.model.ts
interface Geo {
    lat: string;
    lng: string;
}

interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: Geo;
}

interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
}

export interface UserModel {
    id: number;
    name: string;
    username: string;
    email: string;
    profilepicture: string;
    address: Address;
    phone: string;
    website: string;
    company: Company;
}